requirements.txt
beautifulsoup4==4.11.1
matplotlib==3.3.4
nltk==3.6.1
numpy==1.20.1
pandas==1.2.4
Pillow==9.1.1
requests==2.25.1
streamlit==1.3.1
streamlit_authenticator==0.2.1
streamlit_lottie==0.0.3
textblob==0.17.1
wordcloud==1.8.1
